datapath-windows: Fix issues around conntrack
After SET action with the attribute OVS_KEY_ATTR_TUNNEL we need to update
the flow key to accommodate for new offsets which should be used.

Re-extract the flow key before sending it to the conntracking module.

Increase size of the tunnel key attributes.

+++ datapath-windows/ovsext/Flow.c
@@ -2418,7 +2418,7 @@ OvsExtractFlow(const NET_BUFFER_LIST *packet,
 
     if (tunKey) {
         ASSERT(!IsNullIpAddr(&tunKey->dst));
-        UINT8 optOffset = TunnelKeyGetOptionsOffset(tunKey);
+        UINT16 optOffset = TunnelKeyGetOptionsOffset(tunKey);
         RtlMoveMemory(((UINT8 *)&flow->tunKey) + optOffset,
                       ((UINT8 *)tunKey) + optOffset,
                       TunnelKeyGetRealSize(tunKey));
@@ -2762,7 +2762,7 @@ OvsLookupFlow(OVS_DATAPATH *datapath,
     UINT16 size = key->l2.keyLen;
     UINT8 *start;
 
-    ASSERT(!IsNullIpAddr(&key->tunKey.dst) || offset == sizeof(OvsTunnelKey ));
+    ASSERT(!IsNullIpAddr(&key->tunKey.dst) || offset == sizeof(OvsTunnelKey));
     ASSERT(IsNullIpAddr(&key->tunKey.dst) || offset == OvsGetFlowL2Offset(&key->tunKey));
 
     start = (UINT8 *)key + offset;
